Update timing of returning connection and server pipelining CTR.
The connection was being returned back to the pool too early as
sometimes the LastHttpContent had not yet been received. Changing it to
only release after the LastHttpContent has been received will prevent
accidentally using the same channel for multiple requests.

The server will now ignore incoming requests if it is already handling a
request which stops pipelining from working altogether. This prevents
accidentally using the same channel for multiple requests on the server
side which caused errors due to concurrent modification of channel
state.