Prevent Java driver channel from being reused too early.
The connection was being returned too quickly before all HttpContent
were read. This modifies the driver's handlers by ensuring that the
LastHttpContent has been handled before returning the connection. The
same was done for the SimpleHttpClient. As a consequence, some tests
needed to be updated as they will now have up to two more empty
ResponseMessageV4s added to the end.